+++ c++/test/run_info_table_tests.cpp
@@ -36,7 +36,7 @@ SCENARIO("Run Info table Tests")
         auto run_info_data_1 = get_test_run_info_data("_2");
 
         {
-            auto file_out = std::make_shared<pod5::AsyncOutputStream>(
+            auto file_out = *pod5::AsyncOutputStream::make(
                 *arrow::io::FileOutputStream::Open(filename), pod5::make_thread_pool(1));
             auto schema_metadata = make_schema_key_value_metadata(
                 {file_identifier, "test_software", *parse_version_number(Pod5Version)});