+++ c++/test/signal_table_tests.cpp
@@ -43,7 +43,7 @@ SCENARIO("Signal table Tests")
         auto signal_type = GENERATE(SignalType::UncompressedSignal, SignalType::VbzSignal);
 
         {
-            auto file_out = std::make_shared<pod5::AsyncOutputStream>(
+            auto file_out = *pod5::AsyncOutputStream::make(
                 *arrow::io::FileOutputStream::Open(filename), pod5::make_thread_pool(1));
             auto schema_metadata = make_schema_key_value_metadata(
                 {file_identifier, "test_software", *parse_version_number(Pod5Version)});