+++ CHANGELOG.md
@@ -7,6 +7,12 @@ All notable changes, updates, and fixes to pod5 will be documented here
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.20]
+
+## Changed
+
+- Refactored directio writing engine to open up async io support.
+
 ## [0.3.19]
 
 ## Added
@@ -26,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Added
 
 - Ability to disable flushing on batch complete
+- Use new LinuxOutputStream to cache allocations and reduce memory when writing many files.
 
 ## [0.3.17]